+++ src/Listbox/Listbox.test.js
@@ -1,5 +1,5 @@
 /* eslint-disable max-len */
-import { Listbox, Popup } from 'root';
+import { Listbox } from 'root';
 import { events } from '../lib/events';
 
 const {
@@ -11,6 +11,8 @@ const {
   keydownDown,
   keydownHome,
   keydownEnd,
+  keyUpUp,
+  keyUpDown,
 } = events;
 
 const listboxMarkup = `
@@ -63,7 +65,7 @@ describe('Listbox with default configuration', () => {
       const [firstListItem] = listItems;
       expect(listbox.getState().activeDescendant).toEqual(firstListItem);
 
-      expect(controller.popup).toBeInstanceOf(Popup);
+      // expect(controller.popup).toBeInstanceOf(Popup);
 
       expect(controller.listbox).toBeInstanceOf(Listbox);
       expect(target.listbox).toBeInstanceOf(Listbox);
@@ -111,17 +113,15 @@ describe('Listbox with default configuration', () => {
     });
 
     it('Should open the popup on controller DOWN arrow key', () => {
-      controller.dispatchEvent(keydownDown);
+      controller.dispatchEvent(keyUpDown);
       expect(document.activeElement).toEqual(target);
-      // @todo Why does this fail?!?!
-      // expect(listbox.getState().expanded).toBeTruthy();
+      expect(listbox.getState().expanded).toBeTruthy();
     });
 
     it('Should open the popup on controller UP arrow key', () => {
-      controller.dispatchEvent(keydownUp);
+      controller.dispatchEvent(keyUpUp);
       expect(document.activeElement).toEqual(target);
-      // @todo Why does this fail?!?!
-      // expect(listbox.getState().expanded).toBeTruthy();
+      expect(listbox.getState().expanded).toBeTruthy();
     });
   });
 
@@ -148,7 +148,7 @@ describe('Listbox with default configuration', () => {
       expect(document.activeElement).toEqual(controller);
     });
 
-    it('Should set next element as activedescendant on target UP arrow key', () => {
+    it('Should set previous element as activedescendant on target UP arrow key', () => {
       listbox.setState({ activeDescendant: target.children[3] });
       expect(target.children[3].getAttribute('aria-selected')).toEqual('true');
 
@@ -160,7 +160,7 @@ describe('Listbox with default configuration', () => {
       expect(target.children[2].getAttribute('aria-selected')).toEqual('true');
     });
 
-    it('Should set previous element as activedescendant on target DOWN arrow key', () => {
+    it('Should set next element as activedescendant on target DOWN arrow key', () => {
       listbox.setState({ activeDescendant: target.children[4] });
       expect(target.children[4].getAttribute('aria-selected')).toEqual('true');